requirements.txt
flask
ultralytics
sort-track